Property-Reserve/subscriber/requirements.txt
paho-mqtt
# python-dotenv
# descomentar en local
requests
